requirements.txt
beautifulsoup4==4.11.1
python-telegram-bot==20.0
